requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cssselect==1.1.0
fake-useragent==0.1.11
h11==0.12.0
idna==2.10
importlib-metadata==4.10.0
lxml==4.7.1
mypy-extensions==0.4.3
numpy==1.22.2
oauthlib==3.1.0
outcome==1.1.0
pandas==1.4.0
parse==1.19.0
pathspec==0.9.0
platformdirs==2.4.1
polygon-api-client==0.2.11
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pyOpenSSL==21.0.0
pyppeteer==1.0.2
PyQt6==6.2.2
PyQt6-Qt6==6.2.2
PyQt6-sip==13.2.0
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
toml==0.10.2
tomli==1.2.3
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.4
w3lib==1.22.0
watchdog==2.1.6
webdriver-manager==3.5.2
websocket-client==1.2.3
websockets==10.1
wsproto==1.0.0
zipp==3.7.0